+++ apps/shop/src/features/cart/hooks/index.ts
@@ -1,4 +1,8 @@
 export * from './mutations/useCartItemMutation';
 export * from './mutations/useDeleteCartMutation';
 export * from './mutations/useUpdateCartMutation';
+export * from './queries/useCartQuery';
+export * from './queries/useCartCountQuery';
 export * from './useCart';
+
+export * from './queries';
