+++ apps/shop/src/features/cart/hooks/queries/useCartCountQuery.ts
@@ -1,0 +1,9 @@
+import { cartQueries } from '@/features/cart';
+import { useQuery } from '@tanstack/react-query';
+
+export function useCartCountQuery(userId: string) {
+  return useQuery({
+    ...cartQueries.count(),
+    enabled: !!userId,
+  });
+}
