+++ apps/shop/src/features/cart/index.ts
@@ -2,4 +2,3 @@ export * from './components';
 export * from './hooks';
 export * from './types';
 export * from './apis';
-export * from './queries';
