+++ apps/shop/src/shared/components/CartNavBtn.tsx
@@ -3,9 +3,9 @@
 import { CartBadge } from '@/shared/components';
 import { useCartBadge } from '@/shared/hooks';
 import Link from 'next/link';
-import styles from './CartNavLink.module.scss';
+import styles from './CartNavBtn.module.scss';
 
-export default function CartNavLink() {
+export default function CartNavBtn() {
   const { badgeCount } = useCartBadge();
 
   return (
