+++ apps/shop/src/shared/components/index.ts
@@ -3,5 +3,5 @@ export { default as QuantityController } from './QuantityController';
 export { default as CheckBox } from './CheckBox';
 export { default as Dropdown } from './Dropdown';
 export { default as NoData } from './NoData';
-export { default as CartNavLink } from './CartNavLink';
+export { default as CartNavBtn } from './CartNavBtn';
 export { CardSkeleton } from './CardSkeleton';
