+++ apps/shop/src/shared/hooks/useCartBadge.ts
@@ -1,10 +1,9 @@
-import { cartQueries } from '@/features/cart';
+import { useCartCountQuery } from '@/features/cart/hooks';
 import { useUser } from '@/features/user/hooks';
-import { useQuery } from '@tanstack/react-query';
 
 export function useCartBadge() {
   const { userId } = useUser();
-  const { data: cartCount } = useQuery(cartQueries.count(userId));
+  const { data: cartCount } = useCartCountQuery(userId);
 
   return { badgeCount: cartCount?.count ?? 0 };
 }
